Initial Commit
SIAlertView.Xamarin is a C#/Mono Xamarin.iOS port of Sumi Interactive's
[SIAlertView](https://github.com/Sumi-Interactive/SIAlertView) for
native iOS. It's basically a hand-rolled UIAlertView.

Not all of the features work completely. The title and message don't
seem to working reliably. Neither is the gradient background, or the
dismissing animations. I hope to get these working sometime in the near
future.